Finally actually working again!
Turns out that in evegfx-cli I'd switched only one of the two calls over
to use the display timings taken from the gameduino firmware rather than
the ones I derived myself from an earlier project. My timings don't work
because the EVE chip can't match the pixel clock rate I was expecting
well enough.

This mode isn't really in spec and so the monitor I use for testing seems
to lose sync every few minutes. But I'll take this for now so I can go
back to working on the rest of it, and then perhaps revisit seeing if
there's a closer-spec 720p timing mode possible.