evegfx: Allow burst writing to the coprocessor buffer
This restores the earlier behavior of leaving a write transaction open for
most of the lifetime of the EVECoprocessor object, and thus when new writes
come in we will usually be able to just write them directly into the
buffer with no further transaction overhead.

We do occasionally need to stop the write transaction for other work, such
as waiting for more space to be available in the buffer, but for the
common case of streaming a display list into the coprocessor once per
frame and then waiting for vsync this should allow streaming the entire
frame's data in during a single transaction, only ending the transaction
to wait for vsync again.