evegfx: Models are empty structs rather than empty enums
It's actually somewhat convenient to be able to use these both as types
and as values, because it can provide a more ergonomic API in situations
where a function that is generic over models doesn't actually have an
argument to explitily state the type. In that case we can now have an
argument that takes the only value of one of the model types and thus
allow the user to use a more "normal" call syntax rather than the
turbofish syntax.