evegfx: Expose coprocessor faults
The coprocessor signals a fault by changing the read pointer to an
unaligned one, which in turn makes the remaining space appear unaligned.
Our waiter can therefore detect that and report a coprocessor fault as
a special kind of error.

For the BT815 and later chips we can also fetch a string error message
from a special part of the memory space, which gives details about
what failed.