evegfx: Allow vertices and rectangles from tuples
When calling graphics-related functions it's usually clear from context
that a pair of numbers is a vertex coordinate, so we'll allow passing in
a 2-tuple in locations where a 2D vertex is expected to allow callers to
write calls that are more concise.

This also includes a similar mechanism for rectangles, with three
different tuple-based representations available, two of which also include
vertices.