evegfx: Use a specialized error type for the main API
Previously we were just exposing the interface's error type directly, but
having an intermediate error type here makes room for us to signal other
types of failures.

For now, the only other failure we can report is a feature being
unsupported. We're not making use of that yet, but it's in preparation for
supporting other models where we likely won't be able to practically
represent all of the differences statically in the type system.